Up to Exercise 8.21
Refactored LoginForm to utilize onCompleted config for useMutation. Recommendations view has been modified to use lazy queries. The fetch policies have been changed to check against the network alongside the cache due to issues encountered with logging in and out.

For future reference, the setContext function for the ApolloClient object in index.js is run every time a mutation or query is executed. Keep this in mind when using tokens.